% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/master.R
\name{glue_xfplates}
\alias{glue_xfplates}
\title{Glueing mulltiple plates from a folder}
\usage{
glue_xfplates(folderpath_seahorse, arg_is_folder)
}
\arguments{
\item{folderpath_seahorse}{the path to a folder where the .xlsx files
are located or a vector of strings pointing to the path of each individual file}

\item{arg_is_folder}{either TRUE or FALSE. When the input is a vector of path strings
use FALSE, is it points to a folder use TRUE}
}
\value{
a nested tibble with all files organized in a row
}
\description{
This function takes a folder path and on the available .xlsx files
the revive_xfplate() function is run and output in one nested tibble.
}
\examples{
c(
    system.file("extdata",
        "20191219_SciRep_PBMCs_donor_A.xlsx",
        package = "seahtrue"
    ),
    system.file("extdata",
        "20191219_SciRep_PBMCs_donor_A.xlsx",
        package = "seahtrue"
    )
) |>
    glue_xfplates(arg_is_folder = FALSE)
}
