% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rate_method.R
\name{method_id}
\alias{method_id}
\title{The method id of a rate calculation method}
\usage{
method_id(method)
}
\arguments{
\item{method}{A rate calculation method (\code{\link{rate_method}}).}
}
\value{
\code{"name@version"}, matching \code{^[a-z0-9-]+@[0-9]+$}.
}
\description{
The method id of a rate calculation method
}
