% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seahtrue-package.R
\docType{package}
\name{seahtrue-package}
\alias{seahtrue}
\alias{seahtrue-package}
\title{seahtrue: Seahtrue revives XF data for structured data analysis}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Seahtrue organizes oxygen consumption and extracellular acidification analysis data from experiments performed on an XF analyzer into structured nested tibbles.This allows for detailed processing of raw data and advanced data visualization and statistics. Seahtrue introduces an open and reproducible way to analyze these XF experiments. It uses file paths to .xlsx files. These .xlsx files are supplied by the userand are generated by the user in the Wave software from Agilent from the assay result files (.asyr). The .xlsx file contains different sheets of important data for the experiment; 1. Assay Information - Details about how the experiment was set up. 2. Rate Data - Information about the OCR and ECAR rates. 3. Raw Data - The original raw data collected during the experiment. 4. Calibration Data - Data related to calibrating the instrument. Seahtrue focuses on getting the specific data needed for analysis. Once this data is extracted, it is prepared for calculations through preprocessing. To make sure everything is accurate, both the initial data and the preprocessed data go through thorough checks.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://vcjdeboer.github.io/seahtrue/}
  \item Report bugs at \url{https://github.com/vcjdeboer/seahtrue/issues}
}

}
\author{
\strong{Maintainer}: Vincent de Boer \email{vincent.deboer@wur.nl} (\href{https://orcid.org/0000-0001-9928-1698}{ORCID})

Authors:
\itemize{
  \item Vincent de Boer \email{vincent.deboer@wur.nl} (\href{https://orcid.org/0000-0001-9928-1698}{ORCID})
  \item Gerwin Smits
  \item Xiang Zhang
}

}
\keyword{internal}
