% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{seahtrue_output_donor_C}
\alias{seahtrue_output_donor_C}
\title{Dataset donor C}
\format{
A data frame with 1 rows and 7 variables:
\describe{
  \item{plate_id}{Barcode from the 'Assay Configuration' sheet from the Seahorse Excel file.}
  \item{filepath_seahorse}{Path to Seahorse Excel file.}
  \item{date}{Date derived from 'Assay Configuration' sheet of the Seahorse Excel file.}
  \item{assay_info}{Date derived from 'Assay Configuration' sheet of the Seahorse Excel file.}
  \item{injection_info}{Dataframe with information from the 'Assay Configuration' sheet from the Seahorse Excel file.}
  \item{raw_data}{Preprocessed raw dataframe from Seahorse 'Raw' sheet.}
  \item{rate_data}{Preprocessed rate data from 'Rate' sheet.}
}
}
\usage{
seahtrue_output_donor_C
}
\description{
A dataset containing output for donor A. This data contains the preprocess 
output that was generated during the run of our Seahtrue pipeline. This data was 
created with generated with input file '20200110_SciRep_PBMCs_donor_C.xlsx', 
located at the inst/extdata directory.
}
\keyword{datasets}
