% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_xfplate.R
\name{calculate_space}
\alias{calculate_space}
\title{Calculate Bioenergetic Indices from OCR and ECAR Data}
\usage{
calculate_space(
  rate,
  param_set_ocr,
  param_set_ecar,
  conversion_model = "mookerjee",
  ug_protein_scaling_factor = 20,
  OCR_var = "J_oxphos",
  ECAR_var = "J_glyco"
)
}
\arguments{
\item{rate}{A tibble from the `rate_data` list of `revive_xfplate()`.}

\item{param_set_ocr}{Named character vector defining OCR injections (e.g. `init_ocr = "m3"`).}

\item{param_set_ecar}{Named character vector defining ECAR injections (e.g. `om_ecar = "m6"`).}

\item{conversion_model}{Either `"mookerjee"` (default) or `"agilent"`.}

\item{ug_protein_scaling_factor}{Protein content per well (in ug, default = 20).}

\item{OCR_var}{Name of column containing OCR values (default = `"J_oxphos"`).}

\item{ECAR_var}{Name of column containing ECAR values (default = `"J_glyco"`).}
}
\value{
A tibble with one row per group and multiple bioenergetic indices.
}
\description{
Computes bioenergetic parameters (OCR/ECAR) including spare capacity, glycolytic index,
and supply flexibility index. Compatible with both oligomycin- and monensin-based protocols.
}
\examples{
 data <- revive_xfplate(
  system.file("extdata", 
  "20191219_SciRep_PBMCs_donor_A.xlsx", package = "seahtrue")
 ) 
calculate_space(
  rate = purrr::pluck(data, "rate_data", 1),
  param_set_ocr = c(init_ocr = "m3", fccp_ocr = "m4", amrot_ocr = "m9", mon_ocr = "m12"),
  param_set_ecar = c(init_ecar = "m3", fccp_ecar = "m4", amrot_ecar = "m9", mon_ecar = "m12"),
  conversion_model = "mookerjee")

}
