% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seahtrue-package.R
\docType{package}
\name{seahtrue-package}
\alias{seahtrue}
\alias{seahtrue-package}
\title{seahtrue: Introducing Seahtrue, a transformative view on cellular processes using OCR and ECAR.}
\description{
Seahtrue version 0.1.0 introduces an open and reproducible way to analyze Seahorse experiments. It uses file paths to Seahorse Excel files. These files are created from the assay result files (.asyr) which are derived from the Agilent Seahorse XF Wave software. They contain different sheets of important data for the experiment: 1. Assay Information: Details about how the experiment was set up. 2. Rate Data: Information about the rates of different cellular processes. 3. Raw Data: The original data collected during the experiment. 4. Calibration Data: Data related to calibrating the instrument. Seahtrue version 0.1.0 focuses on getting the specific data needed for analysis. Once this data is extracted, it's prepared for calculations through preprocessing. To make sure everything is accurate, both the initial data and the preprocessed data go through thorough checks. The upcoming version of Seahtrue will incorporate comprehensive calculations that build upon the preprocessed data, enhancing its analytical capabilities and providing deeper insights into Seahorse experiment results.
}
\keyword{internal}
